refactor: 디렉토리 구조 변경 - (with-sidebar) 디렉토리 삭제 - sidebar컴포넌트 최상위 layout 파일로 이동 - sidebar컴포넌트에서 페이지 이동 시 주소를 비교하는 연산 추가
