--- src/app/(with-sidebar)/page.tsx
@@ -1,7 +1,0 @@
-export default function Home() {
-  return (
-    <div className='bg-white'>
-      <h2>Leev</h2>
-    </div>
-  )
-}
