+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from 'next'
 import '@/styles/globals.css'
 import TanStackProvider from '@/app/providers'
 import localFont from 'next/font/local'
+import Sidebar from '../components/RootLayout/Sidebar'
 
 const pretendard = localFont({
   src: '../assets/fonts/PretendardVariable.woff2',
@@ -22,7 +23,12 @@ export default function RootLayout({
   return (
     <html lang='en'>
       <body className={`${pretendard.className} antialiased`}>
-        <TanStackProvider>{children}</TanStackProvider>
+        <TanStackProvider>
+          <div className='flex h-screen'>
+            <Sidebar />
+            <div className='grow'>{children}</div>
+          </div>
+        </TanStackProvider>
       </body>
     </html>
   )
