+++ src/app/login/page.tsx
@@ -1,0 +1,6 @@
+const LoginPage = () => {
+  // sidebar 노출 유무 테스트용 페이지
+
+  return <div>login page</div>
+}
+export default LoginPage
